# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MacroVanta API
  description: |
    REST API for the MacroVanta app. Handles user authentication with JWT,
    food entry logging, AI-assisted extraction, USDA food search, daily macro
    and micronutrient goals, and nutrition summaries.

    Authentication: obtain a token from `POST /api/auth/login`, then click
    Authorize and paste the token. SwaggerHub adds the `Bearer ` prefix when
    using the bearer auth helper.

    Email verification: accounts must verify their email before logging in.
    Unverified accounts receive a 403 from the login endpoint.

    Dates: all dates are plain strings in `YYYY-MM-DD` format. Summary endpoint
    defaults use the server's current UTC date.

    Development tokens: when the server is not running in production,
    token-generating responses include a `developmentToken` field. Registration
    includes it after creating a verification token. Resend verification includes
    it only when an existing unverified account receives a new verification
    token. Forgot password includes it only when an existing account receives a
    reset token.
  version: 1.1.0
  contact:
    name: Anthony Mainolfi
servers:
  - url: https://dylanwexler.com
    description: Deployed course demonstration server
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/preset-foods/search:
    get:
      tags: [Preset Foods]
      summary: Search USDA FoodData Central
      description: |
        Makes a server-side search request to USDA FoodData Central and maps
        its nutrient data into MacroVanta's food-entry shape. The USDA API key
        stays on the server. Branded results are scaled to a labeled gram or
        milliliter serving when possible; otherwise values remain per 100 g.
      operationId: searchPresetFoods
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
          example: greek yogurt
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 25
      responses:
        "200":
          description: USDA food search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresetFoodSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: User or USDA rate limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: USDA service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/summary/daily:
    get:
      tags: [Summary]
      summary: Daily totals vs. goals
      description: |
        Totals for the given date, the user's goals, and progress toward each
        goal as a percentage rounded to one decimal. If `date` is omitted, the
        server defaults to the current UTC date. Progress is 0 for any goal that
        is unset or zero.
      operationId: getDailySummary
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DateQuery"
      responses:
        "200":
          description: Daily summary
          content:
            application/json:
              schema:
                type: object
                required: [date, totals, goals, progress]
                properties:
                  date:
                    type: string
                    example: "2026-07-07"
                  totals:
                    $ref: "#/components/schemas/MacroTotals"
                  goals:
                    $ref: "#/components/schemas/MacroTotals"
                  progress:
                    allOf:
                      - $ref: "#/components/schemas/MacroTotals"
                      - description: Percent of each goal reached, rounded to one decimal
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/summary/by-meal:
    get:
      tags: [Summary]
      summary: Daily totals grouped by meal
      description: |
        Totals for each meal type on the given date. If `date` is omitted, the
        server defaults to the current UTC date. All four meals are always
        present.
      operationId: getSummaryByMeal
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DateQuery"
      responses:
        "200":
          description: Per-meal summary
          content:
            application/json:
              schema:
                type: object
                required: [date, meals]
                properties:
                  date:
                    type: string
                    example: "2026-07-07"
                  meals:
                    $ref: "#/components/schemas/MealTotals"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/summary/weekly:
    get:
      tags: [Summary]
      summary: Seven-day summary
      description: |
        Daily totals for the 7 days beginning at `startDate`. If `startDate` is
        omitted, the server defaults to the current UTC date.
      operationId: getWeeklySummary
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: false
          description: First day of the 7-day window, in YYYY-MM-DD format. Defaults to the current UTC date.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
          example: "2026-07-01"
      responses:
        "200":
          description: Weekly summary
          content:
            application/json:
              schema:
                type: object
                required: [startDate, endDate, days]
                properties:
                  startDate:
                    type: string
                    example: "2026-07-01"
                  endDate:
                    type: string
                    example: "2026-07-07"
                  days:
                    type: array
                    items:
                      $ref: "#/components/schemas/DailyTotals"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by `POST /api/auth/login`
  parameters:
    DateQuery:
      name: date
      in: query
      required: false
      description: Filter by date, in YYYY-MM-DD format. Summary endpoints default to the current UTC date.
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
      example: "2026-07-07"
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidDate:
              value:
                message: date must use YYYY-MM-DD format.
            validation:
              value:
                message: At least one field is required.
            weakPassword:
              value:
                message: Password must be at least 8 characters and include uppercase, lowercase, number, and special character.
    Unauthorized:
      description: Missing, invalid, or expired authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                message: Authentication token required.
            invalid:
              value:
                message: Invalid authentication token.
            expired:
              value:
                message: Invalid or expired authentication token.
...
  schemas:
...
    DateString:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
      example: "2026-07-07"
...
    PresetFood:
      type: object
      required: [fdcId, foodName, brand, dataType, servingSize, calories, protein, carbs, fat, saturatedFat, transFat, fiber, sodium, potassium, calcium, iron, vitaminC, vitaminD]
      properties:
        fdcId:
          type: integer
          example: 171077
        foodName:
          type: string
          example: Chicken breast, cooked, roasted
        brand:
          type: string
          nullable: true
        dataType:
          type: string
          example: SR Legacy
        servingSize:
          type: string
          example: 100 g
        calories:
          type: number
          example: 165
        protein:
          type: number
          example: 31
        carbs:
          type: number
          example: 0
        fat:
          type: number
          example: 3.6
        saturatedFat:
          type: number
          example: 1
        transFat:
          type: number
          example: 0
        fiber:
          type: number
          example: 0
        sodium:
          type: number
          example: 74
        potassium:
          type: number
          example: 256
        calcium:
          type: number
          example: 15
        iron:
          type: number
          example: 1
        vitaminC:
          type: number
          example: 0
        vitaminD:
          type: number
          example: 0.1
    PresetFoodSearchResponse:
      type: object
      required: [foods, page, totalPages, totalResults]
      properties:
        foods:
          type: array
          items:
            $ref: "#/components/schemas/PresetFood"
        page:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 20
        totalResults:
          type: integer
          example: 197
...
    MacroTotals:
      type: object
      required: [calories, protein, carbs, fat, saturatedFat, transFat, fiber, sodium, potassium, calcium, iron, vitaminC, vitaminD]
      properties:
        calories:
          type: number
          example: 1850
        protein:
          type: number
          example: 140
        carbs:
          type: number
          example: 180
        fat:
          type: number
          example: 62
        saturatedFat:
          type: number
          example: 16
        transFat:
          type: number
          example: 1
        fiber:
          type: number
          example: 24
        sodium:
          type: number
          example: 1900
        potassium:
          type: number
          example: 3500
        calcium:
          type: number
          example: 900
        iron:
          type: number
          example: 14
        vitaminC:
          type: number
          example: 75
        vitaminD:
          type: number
          example: 12
    MealTotals:
      type: object
      required: [Breakfast, Lunch, Dinner, Snack]
      properties:
        Breakfast:
          $ref: "#/components/schemas/MacroTotals"
        Lunch:
          $ref: "#/components/schemas/MacroTotals"
        Dinner:
          $ref: "#/components/schemas/MacroTotals"
        Snack:
          $ref: "#/components/schemas/MacroTotals"
    DailyTotals:
      type: object
      required: [date, totals]
      properties:
        date:
          $ref: "#/components/schemas/DateString"
        totals:
          $ref: "#/components/schemas/MacroTotals"
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        stack:
          type: string
          description: Stack trace, returned only outside production